refactor and variational cca with numpyro
