+++ cca_zoo/deepmodels/dcca_noi.py
@@ -5,7 +5,7 @@ import torch
 from cca_zoo.deepmodels.architectures import BaseEncoder, Encoder
 from cca_zoo.deepmodels.dcca import DCCA
 from cca_zoo.deepmodels.objectives import _compute_matrix_power, CCA
-from cca_zoo.models.wrappers import MCCA
+from cca_zoo.models import MCCA
 
 
 class DCCA_NOI(DCCA, torch.nn.Module):
@@ -14,7 +14,7 @@ class DCCA_NOI(DCCA, torch.nn.Module):
 
     Examples
     --------
-    >>> from cca_zoo.dcca_noi import DCCA_NOI
+    >>> from cca_zoo.deepmodels import DCCA_NOI
     >>> model = DCCA_NOI()
     """
